using Will Saunders' code to get titles from bibtex
